requirements.txt
# Core dependencies
requests>=2.31.0
python-dotenv>=1.0.0
anthropic>=0.40.0

# Database
sqlalchemy>=2.0.0

# Google API
google-auth>=2.23.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.100.0

# Data processing
numpy>=1.24.0
pandas>=2.0.0

# Time handling
pytz>=2023.3
